corrected shape of bernouilli parameter array

+++ global_code/emcat.py
@@ -299,7 +299,7 @@ class KK(object):
         
         clusters_to_kill = []
         
-        bern = zeros(num_clusters, num_KKruns, max_Dk_size)
+        bern = zeros((num_clusters, num_KKruns, max_Dk_size), dtype = float32)
         
         ########### M step ########################################################
         # Normalize by total number of points to give class weight
